make `databases/`, then move `table_config` & `db_utils` there
